daemon: Fix code smell implicit conversion loses integer precision
- ftell returns long, doubt path will be > int32 or 64 in length, should
  be fine to keep int
- Could convert to long, but then same issue in eet_data_undump thus
  ftell output stored as int

+++ src/daemon/entrance_config.c
@@ -66,7 +66,7 @@ _users_get(void)
      }
 
    fseek(f, 0, SEEK_END);
-   textlen = ftell(f);
+   textlen = (int)ftell(f);
    rewind(f);
    if(textlen > 0)
      {
